store/wizard: move compliance types to submodule
Types are the foundation that other modules depend on. Moving
them first establishes the compliance submodule boundary.
