+++ src/store/slices/wizard/index.ts
@@ -3,5 +3,6 @@ export * from './slice';
 export * from './types';
 
 // submodules
+export * from './compliance/';
 export * from './details';
 export * from './filesystem';
